Refactor post review
