+++ .github/workflows/main.yaml
@@ -48,7 +48,7 @@ on:
         description: Runner template to use
         required: true
         type: string
-      downstream_k8s_version:
+      downstream_k8s_minor_version:
         description: Downstream cluster K8s version to test
         type: string
       downstream_cluster_cleanup:
@@ -73,7 +73,7 @@ env:
   QASE_RUN_COMPLETE: 1
   RANCHER_LOG_COLLECTOR: ${{ github.workspace }}/.github/scripts/collect-rancher-logs.sh
   GCP_RUNNER_ZONE: asia-south2-c
-  DOWNSTREAM_KUBERNETES_VERSION: ${{ inputs.downstream_k8s_version }}
+  DOWNSTREAM_K8S_MINOR_VERSION: ${{ inputs.downstream_k8s_minor_version }}
   DOWNSTREAM_CLUSTER_CLEANUP: ${{ inputs.downstream_cluster_cleanup }}
   QASE_HELPER: ${{ github.workspace }}/hosted/helpers/qase/helper_qase.go
 jobs:
@@ -259,7 +259,7 @@ jobs:
           QASE_RUN_ID: ${{ steps.qase.outputs.qase_run_id }}
           KUBECONFIG: /etc/rancher/k3s/k3s.yaml
           RANCHER_VERSION: ${{ env.RANCHER_VERSION }}
-          K8S_UPGRADE_MINOR_VERSION: ${{ env.DOWNSTREAM_KUBERNETES_VERSION }}
+          DOWNSTREAM_K8S_MINOR_VERSION: ${{ env.DOWNSTREAM_K8S_MINOR_VERSION }}
         run: |
           make e2e-k8s-chart-support-provisioning-tests
 
@@ -272,7 +272,7 @@ jobs:
           QASE_RUN_ID: ${{ steps.qase.outputs.qase_run_id }}
           KUBECONFIG: /etc/rancher/k3s/k3s.yaml
           RANCHER_VERSION: ${{ env.RANCHER_VERSION }}
-          K8S_UPGRADE_MINOR_VERSION: ${{ env.DOWNSTREAM_KUBERNETES_VERSION }}
+          DOWNSTREAM_K8S_MINOR_VERSION: ${{ env.DOWNSTREAM_K8S_MINOR_VERSION }}
         run: |
           make e2e-k8s-chart-support-importing-tests
 
@@ -311,7 +311,7 @@ jobs:
           echo "## General information" >> ${GITHUB_STEP_SUMMARY}
           echo "Rancher Manager Version: ${{ inputs.rancher_version }}" >> ${GITHUB_STEP_SUMMARY}
           echo "K3s on Rancher Manager: ${{ inputs.k3s_version }}" >> ${GITHUB_STEP_SUMMARY}
-          echo "Downstream ${{ inputs.hosted_provider }} version: ${{ inputs.downstream_k8s_version }}" >> ${GITHUB_STEP_SUMMARY}
+          echo "Downstream ${{ inputs.hosted_provider }} minor version: ${{ inputs.downstream_k8s_minor_version }}" >> ${GITHUB_STEP_SUMMARY}
 
   delete-runner:
     if: ${{ always() && inputs.destroy_runner == true }}
