+++ hosted/aks/k8s_chart_support/upgrade/k8s_chart_support_provisioning_upgrade_test.go
@@ -1,4 +1,4 @@
-package chart_support_test
+package k8s_chart_support_upgrade_test
 
 import (
 	"fmt"
@@ -13,43 +13,39 @@ import (
 	"github.com/rancher/hosted-providers-e2e/hosted/helpers"
 )
 
-var _ = Describe("K8sChartSupportImport", func() {
-	var cluster *management.Cluster
-
+var _ = Describe("K8sChartSupportProvisioningUpgrade", func() {
+	var (
+		cluster *management.Cluster
+	)
 	BeforeEach(func() {
 		var err error
-		err = helper.CreateAKSClusterOnAzure(location, clusterName, k8sVersion, "1")
-		Expect(err).To(BeNil())
-
-		aksConfig := new(helper.ImportClusterConfig)
+		aksConfig := new(aks.ClusterConfig)
 		config.LoadAndUpdateConfig(aks.AKSClusterConfigConfigurationFileKey, aksConfig, func() {
 			aksConfig.ResourceGroup = clusterName
+			dnsPrefix := clusterName + "-dns"
+			aksConfig.DNSPrefix = &dnsPrefix
 			aksConfig.ResourceLocation = location
 			aksConfig.Tags = helper.GetTags()
+			aksConfig.KubernetesVersion = &k8sVersion
 		})
-
-		cluster, err = helper.ImportAKSHostedCluster(ctx.RancherClient, clusterName, ctx.CloudCred.ID, false, false, false, false, map[string]string{})
+		cluster, err = aks.CreateAKSHostedCluster(ctx.RancherClient, clusterName, ctx.CloudCred.ID, false, false, false, false, map[string]string{})
 		Expect(err).To(BeNil())
 		cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherClient)
 		Expect(err).To(BeNil())
-		// Workaround to add new Nodegroup till https://github.com/rancher/aks-operator/issues/251 is fixed
-		cluster.AKSConfig = cluster.AKSStatus.UpstreamSpec
 	})
+
 	AfterEach(func() {
 		if ctx.ClusterCleanup {
 			err := helper.DeleteAKSHostCluster(cluster, ctx.RancherClient)
-			Expect(err).To(BeNil())
-			err = helper.DeleteAKSClusteronAzure(clusterName)
 			Expect(err).To(BeNil())
 		} else {
 			fmt.Println("Skipping downstream cluster deletion: ", clusterName)
 		}
 	})
 
-	It(fmt.Sprintf("should successfully test k8s %s chart support importing on upgraded rancher %s", helpers.K8sUpgradedMinorVersion, helpers.RancherVersion), func() {
-		testCaseID = 323 // Report to Qase
-		commonchecks(&ctx, cluster)
-
+	It(fmt.Sprintf("should successfully test k8s %s chart support on rancher %s", helpers.K8sUpgradedMinorVersion, helpers.RancherUpgradeVersion), func() {
+		testCaseID = 320 // Report to Qase
+		commonchecks(&ctx, cluster, clusterName, helpers.RancherUpgradeVersion, helpers.RancherHostname, helpers.K8sUpgradedMinorVersion)
 	})
 
 })
