+++ hosted/eks/k8s_chart_support/k8s_chart_support_import_test.go
@@ -1,4 +1,4 @@
-package chart_support_upgrade_test
+package k8s_chart_support_test
 
 import (
 	"fmt"
@@ -13,7 +13,7 @@ import (
 	"github.com/rancher/hosted-providers-e2e/hosted/helpers"
 )
 
-var _ = Describe("K8sChartSupportImportUpgrade", func() {
+var _ = Describe("K8sChartSupportImport", func() {
 	var cluster *management.Cluster
 	BeforeEach(func() {
 		var err error
@@ -31,7 +31,7 @@ var _ = Describe("K8sChartSupportImportUpgrade", func() {
 		Expect(err).To(BeNil())
 		cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherClient)
 		Expect(err).To(BeNil())
-		//Workaround to add new Nodegroup till https://github.com/rancher/aks-operator/issues/251 is fixed
+		// Workaround to add new Nodegroup till https://github.com/rancher/aks-operator/issues/251 is fixed
 		cluster.EKSConfig = cluster.EKSStatus.UpstreamSpec
 
 	})
@@ -41,14 +41,13 @@ var _ = Describe("K8sChartSupportImportUpgrade", func() {
 			Expect(err).To(BeNil())
 			err = helper.DeleteEKSClusterOnAWS(region, clusterName)
 			Expect(err).To(BeNil())
-			// TODO: Force delete EKS cluster
 		} else {
 			fmt.Println("Skipping downstream cluster deletion: ", clusterName)
 		}
 	})
-	It(fmt.Sprintf("should successfully test k8s %s chart support on rancher %s", helpers.K8sUpgradedMinorVersion, helpers.RancherUpgradeVersion), func() {
-		testCaseID = 318 // Report to Qase
 
-		commonchecks(&ctx, cluster, clusterName, helpers.RancherUpgradeVersion, helpers.RancherHostname, helpers.K8sUpgradedMinorVersion)
+	It(fmt.Sprintf("should successfully test k8s %s chart support importing on upgraded rancher %s", k8sVersion, helpers.RancherVersion), func() {
+		testCaseID = 319 // Report to Qase
+		commonchecks(&ctx, cluster)
 	})
 })
