+++ hosted/eks/k8s_chart_support/k8s_chart_support_suite_test.go
@@ -1,4 +1,4 @@
-package chart_support_test
+package k8s_chart_support_test
 
 import (
 	"fmt"
@@ -7,10 +7,8 @@ import (
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"github.com/rancher-sandbox/ele-testhelpers/kubectl"
 	"github.com/rancher-sandbox/ele-testhelpers/tools"
 	. "github.com/rancher-sandbox/qase-ginkgo"
-	"github.com/rancher/shepherd/clients/rancher/catalog"
 	management "github.com/rancher/shepherd/clients/rancher/generated/management/v3"
 	"github.com/rancher/shepherd/extensions/clusters"
 	namegen "github.com/rancher/shepherd/pkg/namegenerator"
@@ -35,10 +33,7 @@ var _ = BeforeSuite(func() {
 	Expect(helpers.Kubeconfig).ToNot(BeEmpty())
 
 	By("Adding the necessary chart repos", func() {
-		err := kubectl.RunHelmBinaryWithCustomErr("repo", "add", catalog.RancherChartRepo, "https://charts.rancher.io")
-		Expect(err).To(BeNil())
-		err = kubectl.RunHelmBinaryWithCustomErr("repo", "add", fmt.Sprintf("rancher-%s", helpers.RancherChannel), fmt.Sprintf("https://releases.rancher.com/server-charts/%s", helpers.RancherChannel))
-		Expect(err).To(BeNil())
+		helpers.AddRancherCharts()
 	})
 
 })
@@ -60,12 +55,7 @@ var _ = BeforeEach(func() {
 var _ = AfterEach(func() {
 
 	By("Uninstalling the existing operator charts", func() {
-		charts := helpers.ListOperatorChart()
-		for _, chart := range charts {
-			args := []string{"uninstall", chart.Name, "--namespace", helpers.CattleSystemNS}
-			err := kubectl.RunHelmBinaryWithCustomErr(args...)
-			Expect(err).To(BeNil())
-		}
+		helpers.UninstallProviderCharts()
 	})
 })
 
@@ -83,39 +73,20 @@ func commonchecks(ctx *helpers.Context, cluster *management.Cluster) {
 	var originalChartVersion string
 
 	By("checking the chart version", func() {
-		oldCharts := helpers.ListOperatorChart()
-		originalChartVersion = oldCharts[0].DerivedVersion
+		originalChartVersion = helpers.GetCurrentChartVersion()
 		Expect(originalChartVersion).ToNot(BeEmpty())
 		GinkgoLogr.Info("Original chart version: " + originalChartVersion)
 	})
 
 	var downgradedVersion string
 	By("obtaining a version to downgrade", func() {
-		chart := helpers.ListOperatorChart()[0]
-		chartVersions := helpers.ListChartVersions(chart.Name)
-		for _, chartVersion := range chartVersions {
-			if helpers.VersionCompare(chartVersion.DerivedVersion, originalChartVersion) == -1 {
-				downgradedVersion = chartVersion.DerivedVersion
-				break
-			}
-		}
+		downgradedVersion = helpers.GetDowngradeChartVersion(originalChartVersion)
 		Expect(downgradedVersion).ToNot(BeEmpty())
 		GinkgoLogr.Info("Downgrading to version: " + downgradedVersion)
 	})
 
 	By("downgrading the chart version", func() {
-		newCharts := helpers.ListOperatorChart()
-		for _, chart := range newCharts {
-			err := kubectl.RunHelmBinaryWithCustomErr("upgrade", "--install", chart.Name, fmt.Sprintf("%s/%s", catalog.RancherChartRepo, chart.Name), "--namespace", helpers.CattleSystemNS, "--version", downgradedVersion, "--wait")
-			Expect(err).To(BeNil())
-		}
-		// wait until the downgraded chart version is same as the old version
-		Eventually(func() int {
-			charts := helpers.ListOperatorChart()
-			downgradedChartVersion := charts[0].DerivedVersion
-			return helpers.VersionCompare(downgradedChartVersion, originalChartVersion)
-		}, tools.SetTimeout(1*time.Minute), 5*time.Second).Should(BeNumerically("==", -1))
-
+		helpers.DowngradeProviderChart(downgradedVersion)
 	})
 
 	initialNodeCount := *cluster.EKSConfig.NodeGroups[0].DesiredSize
@@ -133,29 +104,16 @@ func commonchecks(ctx *helpers.Context, cluster *management.Cluster) {
 	})
 
 	By("uninstalling the operator chart", func() {
-		charts := helpers.ListOperatorChart()
-		for _, chart := range charts {
-			args := []string{"uninstall", chart.Name, "--namespace", helpers.CattleSystemNS}
-			err := kubectl.RunHelmBinaryWithCustomErr(args...)
-			Expect(err).To(BeNil())
-		}
+		helpers.UninstallProviderCharts()
 	})
 
-	By("making a change(scaling nodegroup down) to the cluster to re-install the operator and validating it is re-installed to the latest version", func() {
+	By("making a change(scaling nodegroup down) to the cluster to re-install the operator and validating it is re-installed to the latest/original version", func() {
 		var err error
 		cluster, err = helper.ScaleNodeGroup(cluster, ctx.RancherClient, initialNodeCount)
 		Expect(err).To(BeNil())
 
-		By("ensuring that the chart is re-installed to the latest version", func() {
-			Eventually(func() int {
-				charts := helpers.ListOperatorChart()
-				if len(charts) == 0 {
-					return 10
-				}
-				reinstalledChartVersion := charts[0].DerivedVersion
-				GinkgoLogr.Info("Checking chart reinstallation", reinstalledChartVersion, originalChartVersion, "\n")
-				return helpers.VersionCompare(reinstalledChartVersion, originalChartVersion)
-			}, tools.SetTimeout(1*time.Minute), 1*time.Second).Should(BeNumerically("==", 0))
+		By("ensuring that the chart is re-installed to the latest/original version", func() {
+			helpers.WaitProviderChartInstallation(originalChartVersion, 0)
 		})
 
 		// We do not use WaitClusterToBeUpgraded because it has been flaky here and times out
