+++ hosted/gke/k8s_chart_support/k8s_chart_support_import_test.go
@@ -1,4 +1,4 @@
-package chart_support_upgrade_test
+package k8s_chart_support_test
 
 import (
 	"fmt"
@@ -13,43 +13,48 @@ import (
 	"github.com/rancher/hosted-providers-e2e/hosted/helpers"
 )
 
-var _ = Describe("K8sChartSupportProvisioningUpgrade", func() {
-
+var _ = Describe("K8sChartSupportImport", func() {
 	var (
 		cluster *management.Cluster
 	)
+
 	BeforeEach(func() {
-		gkeConfig := new(management.GKEClusterConfigSpec)
+		err := helper.CreateGKEClusterOnGCloud(zone, clusterName, project, k8sVersion)
+		Expect(err).To(BeNil())
+
+		gkeConfig := new(helper.ImportClusterConfig)
 		config.LoadAndUpdateConfig(gke.GKEClusterConfigConfigurationFileKey, gkeConfig, func() {
 			gkeConfig.ProjectID = project
 			gkeConfig.Zone = zone
 			labels := helper.GetLabels()
 			gkeConfig.Labels = &labels
-			gkeConfig.KubernetesVersion = &k8sVersion
 			for _, np := range gkeConfig.NodePools {
 				np.Version = &k8sVersion
 			}
 		})
-
-		var err error
-		cluster, err = gke.CreateGKEHostedCluster(ctx.RancherClient, clusterName, ctx.CloudCred.ID, false, false, false, false, map[string]string{})
+		cluster, err = helper.ImportGKEHostedCluster(ctx.RancherClient, clusterName, ctx.CloudCred.ID, false, false, false, false, map[string]string{})
 		Expect(err).To(BeNil())
 		cluster, err = helpers.WaitUntilClusterIsReady(cluster, ctx.RancherClient)
 		Expect(err).To(BeNil())
+
+		// Workaround to add new Nodegroup till https://github.com/rancher/aks-operator/issues/251 is fixed
+		cluster.GKEConfig = cluster.GKEStatus.UpstreamSpec
 	})
 
 	AfterEach(func() {
 		if ctx.ClusterCleanup {
 			err := helper.DeleteGKEHostCluster(cluster, ctx.RancherClient)
 			Expect(err).To(BeNil())
+			err = helper.DeleteGKEClusterOnGCloud(zone, project, clusterName)
+			Expect(err).To(BeNil())
+
 		} else {
 			fmt.Println("Skipping downstream cluster deletion: ", clusterName)
 		}
 	})
 
-	It(fmt.Sprintf("should successfully test k8s %s chart support on rancher %s", helpers.K8sUpgradedMinorVersion, helpers.RancherUpgradeVersion), func() {
-		testCaseID = 312 // Report to Qase
-		commonChartSupportUpgrade(&ctx, cluster, clusterName, helpers.RancherUpgradeVersion, helpers.RancherHostname, helpers.K8sUpgradedMinorVersion)
+	It(fmt.Sprintf("should successfully test k8s %s chart support importing on rancher %s", k8sVersion, helpers.RancherVersion), func() {
+		testCaseID = 315 // Report to Qase
+		commonChartSupport(&ctx, cluster)
 	})
-
 })
