+++ hosted/gke/k8s_chart_support/k8s_chart_support_suite_test.go
@@ -1,18 +1,14 @@
-package chart_support_test
+package k8s_chart_support_test
 
 import (
 	"fmt"
 	"testing"
-	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	namegen "github.com/rancher/shepherd/pkg/namegenerator"
 
-	"github.com/rancher-sandbox/ele-testhelpers/kubectl"
-	"github.com/rancher-sandbox/ele-testhelpers/tools"
 	. "github.com/rancher-sandbox/qase-ginkgo"
-	"github.com/rancher/shepherd/clients/rancher/catalog"
 	management "github.com/rancher/shepherd/clients/rancher/generated/management/v3"
 	"github.com/rancher/shepherd/extensions/clusters"
 
@@ -37,10 +33,7 @@ var _ = BeforeSuite(func() {
 	Expect(helpers.Kubeconfig).ToNot(BeEmpty())
 
 	By("Adding the necessary chart repos", func() {
-		err := kubectl.RunHelmBinaryWithCustomErr("repo", "add", catalog.RancherChartRepo, "https://charts.rancher.io")
-		Expect(err).To(BeNil())
-		err = kubectl.RunHelmBinaryWithCustomErr("repo", "add", fmt.Sprintf("rancher-%s", helpers.RancherChannel), fmt.Sprintf("https://releases.rancher.com/server-charts/%s", helpers.RancherChannel))
-		Expect(err).To(BeNil())
+		helpers.AddRancherCharts()
 	})
 
 })
@@ -58,12 +51,7 @@ var _ = BeforeEach(func() {
 
 var _ = AfterEach(func() {
 	By("Uninstalling the existing operator charts", func() {
-		charts := helpers.ListOperatorChart()
-		for _, chart := range charts {
-			args := []string{"uninstall", chart.Name, "--namespace", helpers.CattleSystemNS}
-			err := kubectl.RunHelmBinaryWithCustomErr(args...)
-			Expect(err).To(BeNil())
-		}
+		helpers.UninstallProviderCharts()
 	})
 })
 
@@ -81,39 +69,20 @@ var _ = ReportAfterEach(func(report SpecReport) {
 func commonChartSupport(ctx *helpers.Context, cluster *management.Cluster) {
 	var originalChartVersion string
 	By("checking the chart version", func() {
-		oldCharts := helpers.ListOperatorChart()
-		originalChartVersion = oldCharts[0].DerivedVersion
+		originalChartVersion = helpers.GetCurrentChartVersion()
 		Expect(originalChartVersion).ToNot(BeEmpty())
 		GinkgoLogr.Info("Original chart version: " + originalChartVersion)
 	})
 
 	var downgradedVersion string
 	By("obtaining a version to downgrade", func() {
-		chart := helpers.ListOperatorChart()[0]
-		chartVersions := helpers.ListChartVersions(chart.Name)
-		for _, chartVersion := range chartVersions {
-			if helpers.VersionCompare(chartVersion.DerivedVersion, originalChartVersion) == -1 {
-				downgradedVersion = chartVersion.DerivedVersion
-				break
-			}
-		}
+		downgradedVersion = helpers.GetDowngradeChartVersion(originalChartVersion)
 		Expect(downgradedVersion).ToNot(BeEmpty())
 		GinkgoLogr.Info("Downgrading to version: " + downgradedVersion)
-
 	})
 
 	By("downgrading the chart version", func() {
-		newCharts := helpers.ListOperatorChart()
-		for _, chart := range newCharts {
-			err := kubectl.RunHelmBinaryWithCustomErr("upgrade", "--install", chart.Name, fmt.Sprintf("%s/%s", catalog.RancherChartRepo, chart.Name), "--namespace", helpers.CattleSystemNS, "--version", downgradedVersion, "--wait")
-			Expect(err).To(BeNil())
-		}
-		Eventually(func() int {
-			charts := helpers.ListOperatorChart()
-			downgradedChartVersion := charts[0].DerivedVersion
-			return helpers.VersionCompare(downgradedChartVersion, originalChartVersion)
-		}, tools.SetTimeout(1*time.Minute), 5*time.Second).Should(BeNumerically("==", -1))
-
+		helpers.DowngradeProviderChart(downgradedVersion)
 	})
 
 	By("making a change to the cluster to validate functionality after chart downgrade", func() {
@@ -129,30 +98,17 @@ func commonChartSupport(ctx *helpers.Context, cluster *management.Cluster) {
 	})
 
 	By("uninstalling the operator chart", func() {
-		charts := helpers.ListOperatorChart()
-		for _, chart := range charts {
-			args := []string{"uninstall", chart.Name, "--namespace", helpers.CattleSystemNS}
-			err := kubectl.RunHelmBinaryWithCustomErr(args...)
-			Expect(err).To(BeNil())
-		}
+		helpers.UninstallProviderCharts()
 	})
 
-	By("making a change(adding a nodepool) to the cluster to re-install the operator and validating it is re-installed to the latest version", func() {
+	By("making a change(adding a nodepool) to the cluster to re-install the operator and validating it is re-installed to the latest/original version", func() {
 		currentNodePoolNumber := len(cluster.GKEConfig.NodePools)
 		var err error
 		cluster, err = helper.AddNodePool(cluster, 1, ctx.RancherClient)
 		Expect(err).To(BeNil())
 
-		By("ensuring that the chart is re-installed to the latest version", func() {
-			Eventually(func() int {
-				charts := helpers.ListOperatorChart()
-				if len(charts) == 0 {
-					return 10
-				}
-				reinstalledChartVersion := charts[0].DerivedVersion
-				return helpers.VersionCompare(reinstalledChartVersion, originalChartVersion)
-			}, tools.SetTimeout(1*time.Minute), 5*time.Second).Should(BeNumerically("==", 0))
-
+		By("ensuring that the chart is re-installed to the latest/original version", func() {
+			helpers.WaitProviderChartInstallation(originalChartVersion, 0)
 		})
 
 		err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
