+++ hosted/gke/k8s_chart_support/upgrade/k8s_chart_support_import_upgrade_test.go
@@ -1,4 +1,4 @@
-package chart_support_upgrade_test
+package k8s_chart_support_upgrade_test
 
 import (
 	"fmt"
