+++ hosted/gke/k8s_chart_support/upgrade/k8s_chart_support_upgrade_suite_test.go
@@ -1,4 +1,4 @@
-package chart_support_upgrade_test
+package k8s_chart_support_upgrade_test
 
 import (
 	"fmt"
@@ -7,11 +7,11 @@ import (
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+
 	"github.com/rancher-sandbox/ele-testhelpers/kubectl"
 	"github.com/rancher-sandbox/ele-testhelpers/tools"
 	. "github.com/rancher-sandbox/qase-ginkgo"
 	"github.com/rancher/shepherd/clients/rancher"
-	"github.com/rancher/shepherd/clients/rancher/catalog"
 	management "github.com/rancher/shepherd/clients/rancher/generated/management/v3"
 	"github.com/rancher/shepherd/extensions/clusters"
 	nodestat "github.com/rancher/shepherd/extensions/nodes"
@@ -20,15 +20,16 @@ import (
 	"github.com/rancher/shepherd/pkg/config"
 	namegen "github.com/rancher/shepherd/pkg/namegenerator"
 
-	"github.com/rancher/hosted-providers-e2e/hosted/eks/helper"
+	"github.com/rancher/hosted-providers-e2e/hosted/gke/helper"
 	"github.com/rancher/hosted-providers-e2e/hosted/helpers"
 )
 
 var (
 	ctx                     helpers.Context
 	clusterName, k8sVersion string
-	region                  = helpers.GetEKSRegion()
 	testCaseID              int64
+	zone                    = helpers.GetGKEZone()
+	project                 = helpers.GetGKEProjectID()
 )
 
 func TestK8sChartSupport(t *testing.T) {
@@ -37,19 +38,11 @@ func TestK8sChartSupport(t *testing.T) {
 }
 
 var _ = BeforeSuite(func() {
-	Expect(helpers.RancherVersion).ToNot(BeEmpty())
-	// For upgrade tests, the rancher version should not be an unreleased version (for e.g. 2.8-head)
-	Expect(helpers.RancherVersion).ToNot(ContainSubstring("head"))
-
-	Expect(helpers.RancherUpgradeVersion).ToNot(BeEmpty())
 	Expect(helpers.K8sUpgradedMinorVersion).ToNot(BeEmpty())
 	Expect(helpers.Kubeconfig).ToNot(BeEmpty())
 
 	By("Adding the necessary chart repos", func() {
-		err := kubectl.RunHelmBinaryWithCustomErr("repo", "add", catalog.RancherChartRepo, "https://charts.rancher.io")
-		Expect(err).To(BeNil())
-		err = kubectl.RunHelmBinaryWithCustomErr("repo", "add", fmt.Sprintf("rancher-%s", helpers.RancherChannel), fmt.Sprintf("https://releases.rancher.com/server-charts/%s", helpers.RancherChannel))
-		Expect(err).To(BeNil())
+		helpers.AddRancherCharts()
 	})
 
 	By(fmt.Sprintf("Installing Rancher Manager v%s", helpers.RancherVersion), func() {
@@ -63,28 +56,21 @@ var _ = BeforeEach(func() {
 	ctx, err = helpers.CommonBeforeSuite(helpers.Provider)
 	Expect(err).To(BeNil())
 	clusterName = namegen.AppendRandomString(helpers.ClusterNamePrefix)
-	k8sVersion, err = helper.DefaultEKS(ctx.RancherClient)
+
+	k8sVersion, err = helper.DefaultGKE(ctx.RancherClient, project, ctx.CloudCred.ID, zone, "")
 	Expect(err).To(BeNil())
-	Expect(k8sVersion).ToNot(BeEmpty())
-	GinkgoLogr.Info(fmt.Sprintf("Using EKS version %s", k8sVersion))
+	GinkgoLogr.Info(fmt.Sprintf("Using GKE version %s", k8sVersion))
 
 })
 
 var _ = AfterEach(func() {
 	// The test must restore the env to its original state, so we install rancher back to its original version and uninstall the operator charts
-	// Restoring rancher back to its original state is necessary because in case DOWNSTREAM_CLUSTER_CLEANUP is set to false; in which case clusters will be retained for the next test.
-	// Once the operator is uninstalled, it might be reinstalled since the cluster exists, and installing rancher back to its original state ensures that the version is not the one we want to test.
 	By(fmt.Sprintf("Installing Rancher back to its original version %s", helpers.RancherVersion), func() {
 		helpers.DeployRancherManager(helpers.RancherVersion, true)
 	})
 
 	By("Uninstalling the existing operator charts", func() {
-		charts := helpers.ListOperatorChart()
-		for _, chart := range charts {
-			args := []string{"uninstall", chart.Name, "--namespace", helpers.CattleSystemNS}
-			err := kubectl.RunHelmBinaryWithCustomErr(args...)
-			Expect(err).To(BeNil())
-		}
+		helpers.UninstallProviderCharts()
 	})
 })
 
@@ -98,9 +84,8 @@ var _ = ReportAfterEach(func(report SpecReport) {
 	Qase(testCaseID, report)
 })
 
-func commonchecks(ctx *helpers.Context, cluster *management.Cluster, clusterName, rancherUpgradedVersion, hostname, k8sUpgradedVersion string) {
-	var newChartVersion, oldChartVersion string
-
+// commonChartSupportUpgrade runs the common checks required for testing chart support
+func commonChartSupportUpgrade(ctx *helpers.Context, cluster *management.Cluster, clusterName, rancherUpgradedVersion, hostname, k8sUpgradedVersion string) {
 	By("checking cluster name is same", func() {
 		Expect(cluster.Name).To(BeEquivalentTo(clusterName))
 	})
@@ -121,19 +106,22 @@ func commonchecks(ctx *helpers.Context, cluster *management.Cluster, clusterName
 		Expect(podErrors).To(BeEmpty())
 	})
 
+	var originalChartVersion string
 	By("checking the chart version", func() {
-		oldCharts := helpers.ListOperatorChart()
-		oldChartVersion = oldCharts[0].DerivedVersion
-		Expect(oldChartVersion).ToNot(BeEmpty())
+		originalChartVersion = helpers.GetCurrentChartVersion()
+		Expect(originalChartVersion).ToNot(BeEmpty())
+		GinkgoLogr.Info("Original chart version: " + originalChartVersion)
 	})
 
 	By("upgrading rancher", func() {
 		helpers.DeployRancherManager(rancherUpgradedVersion, true)
+
 		By("ensuring operator pods are also up", func() {
 			Eventually(func() error {
 				return kubectl.New().WaitForNamespaceWithPod(helpers.CattleSystemNS, fmt.Sprintf("ke.cattle.io/operator=%s", helpers.Provider))
 			}, tools.SetTimeout(4*time.Minute), 30*time.Second).Should(BeNil())
 		})
+
 		By("regenerating the token and initiating a new rancher client", func() {
 			//	regenerate the tokens and initiate a new rancher client
 			rancherConfig := new(rancher.Config)
@@ -178,92 +166,65 @@ func commonchecks(ctx *helpers.Context, cluster *management.Cluster, clusterName
 		})
 	})
 
-	By("checking the chart version and validating it is > the old version", func() {
+	var upgradedChartVersion string
 
-		Eventually(func() int {
-			charts := helpers.ListOperatorChart()
-			newChartVersion = charts[0].DerivedVersion
-			return helpers.VersionCompare(newChartVersion, oldChartVersion)
-		}, tools.SetTimeout(1*time.Minute), 5*time.Second).Should(BeNumerically("==", 1))
+	By("checking the chart version and validating it is > the old version", func() {
+		helpers.WaitProviderChartInstallation(originalChartVersion, 1)
+		upgradedChartVersion = helpers.GetCurrentChartVersion()
+		GinkgoLogr.Info("Upgraded chart version: " + upgradedChartVersion)
 
 	})
 
-	var latestVersion *string
-	By(fmt.Sprintf("fetching a list of available k8s versions and ensure the v%s is present in the list and upgrading the cluster to it", k8sUpgradedVersion), func() {
-		versions, err := helper.ListEKSAvailableVersions(ctx.RancherClient, cluster.ID)
+	By(fmt.Sprintf("fetching a list of available k8s versions and ensuring v%s is present in the list and upgrading the cluster to it", k8sUpgradedVersion), func() {
+		versions, err := helper.ListGKEAvailableVersions(ctx.RancherClient, cluster.ID)
 		Expect(err).To(BeNil())
-		Expect(versions).ToNot(BeEmpty())
-		latestVersion = &versions[0]
-		Expect(*latestVersion).To(ContainSubstring(k8sUpgradedVersion))
-		Expect(helpers.VersionCompare(*latestVersion, cluster.Version.GitVersion)).To(BeNumerically("==", 1))
+		latestVersion := versions[len(versions)-1]
+		Expect(latestVersion).To(ContainSubstring(k8sUpgradedVersion))
+		Expect(helpers.VersionCompare(latestVersion, cluster.Version.GitVersion)).To(BeNumerically("==", 1))
 
-		cluster, err = helper.UpgradeClusterKubernetesVersion(cluster, latestVersion, ctx.RancherClient)
+		cluster, err = helper.UpgradeKubernetesVersion(cluster, &latestVersion, ctx.RancherClient, true)
 		Expect(err).To(BeNil())
 		err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
 		Expect(err).To(BeNil())
-		Expect(cluster.EKSConfig.KubernetesVersion).To(BeEquivalentTo(latestVersion))
+		Expect(*cluster.GKEConfig.KubernetesVersion).To(BeEquivalentTo(latestVersion))
+		for _, np := range cluster.GKEConfig.NodePools {
+			Expect(*np.Version).To(BeEquivalentTo(latestVersion))
+		}
 	})
 
 	By("downgrading the chart version", func() {
-		newCharts := helpers.ListOperatorChart()
-		for _, chart := range newCharts {
-			// helm requires 2 args
-			err := kubectl.RunHelmBinaryWithCustomErr("upgrade", "--install", chart.Name, fmt.Sprintf("rancher-charts/%s", chart.Name), "--namespace", helpers.CattleSystemNS, "--version", oldChartVersion, "--wait")
-			Expect(err).To(BeNil())
-		}
-		// wait until the downgraded chart version is same as the old version
-		Eventually(func() int {
-			charts := helpers.ListOperatorChart()
-			downgradedChartVersion := charts[0].DerivedVersion
-			return helpers.VersionCompare(downgradedChartVersion, oldChartVersion)
-		}, tools.SetTimeout(1*time.Minute), 5*time.Second).Should(BeNumerically("==", 0))
-
+		helpers.DowngradeProviderChart(originalChartVersion)
 	})
 
 	By("making a change to the cluster to validate functionality after chart downgrade", func() {
+		initialNodeCount := *cluster.GKEConfig.NodePools[0].InitialNodeCount
 		var err error
-		cluster, err = helper.UpgradeNodeKubernetesVersion(cluster, latestVersion, ctx.RancherClient)
+		cluster, err = helper.ScaleNodePool(cluster, ctx.RancherClient, initialNodeCount+1)
 		Expect(err).To(BeNil())
-		if !cluster.EKSConfig.Imported {
-			// TODO Does not upgrade version for imported cluster, since they use custom Launch Templates
-			err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
-			Expect(err).To(BeNil())
-			for _, ng := range cluster.EKSConfig.NodeGroups {
-				Expect(ng.Version).To(BeEquivalentTo(latestVersion))
-			}
+		err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
+		Expect(err).To(BeNil())
+		for i := range cluster.GKEConfig.NodePools {
+			Expect(*cluster.GKEConfig.NodePools[i].InitialNodeCount).To(BeNumerically(">", initialNodeCount))
 		}
 	})
 
 	By("uninstalling the operator chart", func() {
-		charts := helpers.ListOperatorChart()
-		for _, chart := range charts {
-			args := []string{"uninstall", chart.Name, "--namespace", helpers.CattleSystemNS}
-			err := kubectl.RunHelmBinaryWithCustomErr(args...)
-			Expect(err).To(BeNil())
-		}
+		helpers.UninstallProviderCharts()
 	})
 
-	By("making a change(adding a nodepool) to the cluster to re-install the operator and validating it is re-installed to the latest version", func() {
-		currentNodeGroupNumber := len(cluster.EKSConfig.NodeGroups)
+	By("making a change(adding a nodepool) to the cluster to re-install the operator and validating it is re-installed to the latest/upgraded version", func() {
+		currentNodePoolNumber := len(cluster.GKEConfig.NodePools)
 		var err error
-		cluster, err = helper.AddNodeGroup(cluster, 1, ctx.RancherClient)
+		cluster, err = helper.AddNodePool(cluster, 1, ctx.RancherClient)
 		Expect(err).To(BeNil())
 
-		By("ensuring that the chart is re-installed to the latest version", func() {
-			Eventually(func() int {
-				charts := helpers.ListOperatorChart()
-				if len(charts) == 0 {
-					return 10
-				}
-				reinstalledChartVersion := charts[0].DerivedVersion
-				return helpers.VersionCompare(reinstalledChartVersion, newChartVersion)
-			}, tools.SetTimeout(1*time.Minute), 5*time.Second).Should(BeNumerically("==", 0))
-
+		By("ensuring that the chart is re-installed to the latest/upgraded version", func() {
+			helpers.WaitProviderChartInstallation(upgradedChartVersion, 0)
 		})
 
 		err = clusters.WaitClusterToBeUpgraded(ctx.RancherClient, cluster.ID)
 		Expect(err).To(BeNil())
-		Expect(len(cluster.EKSConfig.NodeGroups)).To(BeNumerically("==", currentNodeGroupNumber+1))
+		Expect(len(cluster.GKEConfig.NodePools)).To(BeNumerically("==", currentNodePoolNumber+1))
 	})
 
 }
